+++ Yandex/src/main/java/com/applovin/mediation/adapters/YandexMediationAdapter.java
@@ -8,8 +8,10 @@ import android.os.Bundle;
 import android.text.TextUtils;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import com.applovin.impl.sdk.utils.BundleUtils;
@@ -83,6 +85,13 @@ public class YandexMediationAdapter
         extends MediationAdapterBase
         implements MaxSignalProvider, MaxInterstitialAdapter, MaxRewardedAdapter, MaxAdViewAdapter /* MaxNativeAdAdapter */
 {
+    private static final int TITLE_LABEL_TAG          = 1;
+    private static final int MEDIA_VIEW_CONTAINER_TAG = 2;
+    private static final int ICON_VIEW_TAG            = 3;
+    private static final int BODY_VIEW_TAG            = 4;
+    private static final int CALL_TO_ACTION_VIEW_TAG  = 5;
+    private static final int ADVERTISER_VIEW_TAG      = 8;
+
     private static final AtomicBoolean INITIALIZED = new AtomicBoolean();
 
     // Required parameters given by Yandex
@@ -159,7 +168,7 @@ public class YandexMediationAdapter
     }
 
     @Override
-    public void initialize(final MaxAdapterInitializationParameters parameters, final Activity activity, final OnCompletionListener onCompletionListener)
+    public void initialize(final MaxAdapterInitializationParameters parameters, @Nullable final Activity activity, final OnCompletionListener onCompletionListener)
     {
         if ( INITIALIZED.compareAndSet( false, true ) )
         {
@@ -199,7 +208,7 @@ public class YandexMediationAdapter
     //region MaxSignalProvider
 
     @Override
-    public void collectSignal(final MaxAdapterSignalCollectionParameters parameters, final Activity activity, final MaxSignalCollectionListener callback)
+    public void collectSignal(final MaxAdapterSignalCollectionParameters parameters, @Nullable final Activity activity, final MaxSignalCollectionListener callback)
     {
         log( "Collecting signal..." );
 
@@ -207,7 +216,7 @@ public class YandexMediationAdapter
 
         BidderTokenRequestConfiguration bidderTokenRequest = createBidderTokenRequestConfiguration( activity, parameters.getAdFormat() );
 
-        BidderTokenLoader.loadBidderToken( activity, bidderTokenRequest, new BidderTokenLoadListener()
+        BidderTokenLoader.loadBidderToken( getContext( activity ), bidderTokenRequest, new BidderTokenLoadListener()
         {
             @Override
             public void onBidderTokenLoaded(@NonNull final String bidderToken)
@@ -232,7 +241,7 @@ public class YandexMediationAdapter
     //region MaxInterstitialAdapter
 
     @Override
-    public void loadInterstitialAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxInterstitialAdapterListener listener)
+    public void loadInterstitialAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxInterstitialAdapterListener listener)
     {
         final String placementId = parameters.getThirdPartyAdPlacementId();
         log( "Loading " + ( AppLovinSdkUtils.isValidString( parameters.getBidResponse() ) ? "bidding " : "" ) + "interstitial ad for placement: " + placementId + "..." );
@@ -266,7 +275,7 @@ public class YandexMediationAdapter
     }
 
     @Override
-    public void showInterstitialAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxInterstitialAdapterListener listener)
+    public void showInterstitialAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxInterstitialAdapterListener listener)
     {
         log( "Showing interstitial ad..." );
 
@@ -274,6 +283,13 @@ public class YandexMediationAdapter
         {
             log( "Interstitial ad failed to load - ad not ready" );
             listener.onInterstitialAdDisplayFailed( new MaxAdapterError( -4205, "Ad Display Failed", 0, "Interstitial ad not ready" ) );
+            return;
+        }
+
+        if ( activity == null )
+        {
+            log( "Interstitial ad display failed: Activity is null" );
+            listener.onInterstitialAdDisplayFailed( MaxAdapterError.MISSING_ACTIVITY );
             return;
         }
 
@@ -286,7 +302,7 @@ public class YandexMediationAdapter
     //region MaxRewardedAdapter
 
     @Override
-    public void loadRewardedAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxRewardedAdapterListener listener)
+    public void loadRewardedAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxRewardedAdapterListener listener)
     {
         final String placementId = parameters.getThirdPartyAdPlacementId();
         log( "Loading " + ( AppLovinSdkUtils.isValidString( parameters.getBidResponse() ) ? "bidding " : "" ) + "rewarded ad for placement: " + placementId + "..." );
@@ -320,7 +336,7 @@ public class YandexMediationAdapter
     }
 
     @Override
-    public void showRewardedAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxRewardedAdapterListener listener)
+    public void showRewardedAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxRewardedAdapterListener listener)
     {
         log( "Showing rewarded ad..." );
 
@@ -328,6 +344,13 @@ public class YandexMediationAdapter
         {
             log( "Rewarded ad failed to load - ad not ready" );
             listener.onRewardedAdDisplayFailed( new MaxAdapterError( -4205, "Ad Display Failed", 0, "Rewarded ad not ready" ) );
+            return;
+        }
+
+        if ( activity == null )
+        {
+            log( "Rewarded ad display failed: Activity is null" );
+            listener.onRewardedAdDisplayFailed( MaxAdapterError.MISSING_ACTIVITY );
             return;
         }
 
@@ -343,7 +366,7 @@ public class YandexMediationAdapter
     //region MaxAdViewAdapter
 
     @Override
-    public void loadAdViewAd(final MaxAdapterResponseParameters parameters, final MaxAdFormat adFormat, final Activity activity, final MaxAdViewAdapterListener listener)
+    public void loadAdViewAd(final MaxAdapterResponseParameters parameters, final MaxAdFormat adFormat, @Nullable final Activity activity, final MaxAdViewAdapterListener listener)
     {
         final String adFormatLabel = adFormat.getLabel();
         final String placementId = parameters.getThirdPartyAdPlacementId();
@@ -374,7 +397,7 @@ public class YandexMediationAdapter
     //region MaxNativeAdapter
 
     @Override
-    public void loadNativeAd(final MaxAdapterResponseParameters parameters, final Activity activity, final MaxNativeAdAdapterListener listener)
+    public void loadNativeAd(final MaxAdapterResponseParameters parameters, @Nullable final Activity activity, final MaxNativeAdAdapterListener listener)
     {
         final String placementId = parameters.getThirdPartyAdPlacementId();
         final String bidResponse = parameters.getBidResponse();
@@ -402,7 +425,7 @@ public class YandexMediationAdapter
 
     //region Helper Methods
 
-    private Context getContext(final Activity activity)
+    private Context getContext(@Nullable final Activity activity)
     {
         // NOTE: `activity` can only be null in 11.1.0+, and `getApplicationContext()` is introduced in 11.1.0
         return ( activity != null ) ? activity.getApplication() : getApplicationContext();
@@ -906,14 +929,17 @@ public class YandexMediationAdapter
 
             nativeAdView = new NativeAdView( container.getContext() );
 
-            // The Yandex Native Ad View needs to be wrapped around the main native ad view to get impressions.
-            MaxNativeAdView maxNativeAdView = (MaxNativeAdView) container;
-            View mainView = maxNativeAdView.getMainView();
-            maxNativeAdView.removeView( mainView );
-            nativeAdView.addView( mainView );
-            maxNativeAdView.addView( nativeAdView );
+            // Native integrations
+            if ( container instanceof MaxNativeAdView )
+            {
+                // The Yandex Native Ad View needs to be wrapped around the main native ad view to get impressions.
+                MaxNativeAdView maxNativeAdView = (MaxNativeAdView) container;
+                View mainView = maxNativeAdView.getMainView();
+                maxNativeAdView.removeView( mainView );
+                nativeAdView.addView( mainView );
+                maxNativeAdView.addView( nativeAdView );
 
-            final NativeAdViewBinder binder = new NativeAdViewBinder.Builder( nativeAdView )
+                final NativeAdViewBinder binder = new NativeAdViewBinder.Builder( nativeAdView )
                     .setIconView( maxNativeAdView.getIconImageView() )
                     .setTitleView( maxNativeAdView.getTitleTextView() )
                     .setDomainView( maxNativeAdView.getAdvertiserTextView() )
@@ -923,17 +949,127 @@ public class YandexMediationAdapter
                     .setCallToActionView( maxNativeAdView.getCallToActionButton() )
                     .build();
 
-            try
-            {
-                YandexMediationAdapter.this.nativeAd.bindNativeAd( binder );
-                nativeAdView.setVisibility( View.VISIBLE );
+                try
+                {
+                    YandexMediationAdapter.this.nativeAd.bindNativeAd( binder );
+                    nativeAdView.setVisibility( View.VISIBLE );
+                }
+                catch ( NativeAdException exception )
+                {
+                    e( "Failed to register native ad views.", exception );
+                }
             }
-            catch ( NativeAdException exception )
+            // Plugins
+            else
             {
-                e( "Failed to register native ad views.", exception );
+                // The Yandex Native Ad View needs to be wrapped around the main native ad view to get impressions.
+                try
+                {
+                    // Remove mediaView from its parent
+                    MediaView mediaView = (MediaView) getMediaView();
+                    ViewGroup mediaViewContainer = (ViewGroup) mediaView.getParent();
+                    mediaViewContainer.removeView( mediaView );
+
+                    // Add mediaView to nativeAdView
+                    ViewGroup.LayoutParams mediaViewLayout = new ViewGroup.LayoutParams( ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT );
+                    nativeAdView.addView( mediaView, mediaViewLayout );
+
+                    // Add nativeAdview to the original parent of the mediaView, noting that Flutter has a layout to add it but React Native doesn't.
+                    boolean hasPluginLayout = ( mediaViewContainer instanceof RelativeLayout || mediaViewContainer instanceof FrameLayout );
+                    if ( hasPluginLayout )
+                    {
+                        ViewGroup.LayoutParams nativeAdViewLayout = new ViewGroup.LayoutParams( ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT );
+                        mediaViewContainer.addView( nativeAdView, nativeAdViewLayout );
+                    }
+                    else
+                    {
+                        nativeAdView.measure(
+                            View.MeasureSpec.makeMeasureSpec( mediaViewContainer.getWidth(), View.MeasureSpec.EXACTLY ),
+                            View.MeasureSpec.makeMeasureSpec( mediaViewContainer.getHeight(), View.MeasureSpec.EXACTLY ) );
+                        nativeAdView.layout( 0, 0, mediaViewContainer.getWidth(), mediaViewContainer.getHeight() );
+                        mediaViewContainer.addView( nativeAdView );
+                    }
+
+                    // Provide the asset views to NativeAdViewBinder for binding them to the Yandex nativeAd.
+                    final NativeAdViewBinder.Builder binder = new NativeAdViewBinder.Builder( nativeAdView );
+                    binder.setMediaView( mediaView );
+
+                    // Set the rest of the asset views from clickableViews.
+                    for ( View view : clickableViews )
+                    {
+                        Object viewTag = view.getTag();
+                        if ( viewTag == null ) continue;
+
+                        int tag = (int) viewTag;
+
+                        if ( tag == ICON_VIEW_TAG )
+                        {
+                            if ( view instanceof ImageView )
+                            {
+                                binder.setIconView( (ImageView) view );
+                            }
+                        }
+                        else if ( tag == TITLE_LABEL_TAG )
+                        {
+                            TextView textView = createTextViewIfNeeded( view );
+                            if ( textView != null )
+                            {
+                                binder.setTitleView( textView );
+                            }
+                        }
+                        else if ( tag == ADVERTISER_VIEW_TAG )
+                        {
+                            TextView textView = createTextViewIfNeeded( view );
+                            if ( textView != null )
+                            {
+                                binder.setDomainView( textView );
+                            }
+                        }
+                        else if ( tag == BODY_VIEW_TAG )
+                        {
+                            TextView textView = createTextViewIfNeeded( view );
+                            if ( textView != null )
+                            {
+                                binder.setBodyView( textView );
+                            }
+                        }
+                        else if ( tag == CALL_TO_ACTION_VIEW_TAG )
+                        {
+                            TextView textView = createTextViewIfNeeded( view );
+                            if ( textView != null )
+                            {
+                                binder.setCallToActionView( textView );
+                            }
+                        }
+                    }
+
+                    YandexMediationAdapter.this.nativeAd.bindNativeAd( binder.build() );
+                    nativeAdView.setVisibility( View.VISIBLE );
+                }
+                catch ( Throwable th )
+                {
+                    e( "Failed to register native ad views.", th );
+                }
             }
 
             return true;
+        }
+
+        private TextView createTextViewIfNeeded(final View view)
+        {
+            TextView textView = null;
+            if ( view instanceof TextView )
+            {
+                textView = (TextView) view;
+            }
+            else if ( view instanceof ViewGroup )
+            {
+                textView = new TextView( view.getContext() );
+                // Hide the text content since this textView is for receiving touch events.
+                textView.setAlpha( 0 );
+                ( (ViewGroup) view ).addView( textView );
+            }
+            return textView;
         }
     }
 }